// requests-service/src/domain/enum/responseStatusCode.enum.ts
export enum ResponseStatusCode {
  //General
  SUCCESS = 2000,
  FAIL = 4000,
  GENERIC_ERROR_500 = 500,

  // SUBMIT
  SUBMIT_FAILURE = 6002,

  //Policy Docs
  NO_POLICY_DOCS = 3020,
  SOME_POLICY_DOCS = 3021,

  //Loan Statements
  NO_ACTIVE_LOANS = 6003,

  //LOAN STATEMENT DOWNLOAD
  NO_STATEMENT_DOWNLOAD = 3031,

  //IP Blacklisting
  FAIL_COUNTRY_IP_BLACKLISTING = 3099,

  //Loan repay
  LOAN_REPAYMENT_SUCCESS = 6040,
  REPAYMENT_STATUS_PENDING = 6999,
  REPAYMENT_STATUS_REJECTED = 6998,
  REPAYMENT_STATUS_FAILURE = 6997,
}